[FIX] sale_order_dates: commitment date for policy all at once
When the policy of the sale order is set to
`Deliver all products at once`,
the commitment date should be the maximum of the order lines lead time
rather than the minimum, as you will deliver all lines at the same time

It looks like the former behavior is applied since Odoo 7.0
d8d0b8846749b5f34f9e29463fd1e1d6e151b91b

It should therefore actually be applied in 7.0/8.0/9.0,
but we consider the risk too high for databases
that are used to the former behavior, and expect the old behavior.

We therefore apply this revision only in 10.0, to minimize the risks.

opw-702271