[FIX] website_sale: star rating for first comment
For first comments on a product in the ecommerce we would not
have the star rating displayed till a page refresh, and the star widget
would not be reset after posting.

Also in other situations, we would get a 0 star widget when posting
comment without rating, but after a refresh the widget would dispappear.

With this change the new comment get the same stars as when refreshed.

opw-702608