[FIX] hr_holidays: add followers to employee leave
from odoo#14848

Those subscribed to the department of the employee didn't get notified when
the employee encoded a leave.

It was due to the department_id not being given as value in create when the
employee made the leave.

This is because there's a group group_hr_holidays_user on that field in the
view, so only when they created a leave the followers would be added correctly
(as department_id was given as value in create).

This commit also remove the useless subscribe_followers method.