[FIX] base: revert and improve 24aba601
It forced to select a coutnry before the state which is not very userfriendly.
Cf comments at #13657

Instead restrict only if the country if a country is selected.
Courtesy of Pedro M. Baeza, Tecnativa
For v10 this is a minimum change and a larger refactoring can be done in master.

This replaces and closes #14896