[FIX] calendar: partner filter for search more or created
When adding a partner "filter" to the calendar view, it only works if we
select directly the partner and not when we choose him via "search more"
or create him on the fly.

This commit uses the partner id directly which is available in the field
value in those three possible instances.

opw-702491