[FIX] stock: use avg for cost in group by report
When doing pivot table data visualization, the Unit Cost of several
lines are summed up but there has more meaning if they were averaged.

This will probably still not be perfect (such as averaging two line
with different quantities the quantities would not be taken into
accounts) but it is already slightly better.

opw-694068