[FIX] mail: action button translation
- Install a language other than English, set it to your user.
- On any model, send a message  (the recipient should be a user, not a
  simple portal user).
- The recipient receives a notification with a primary button half
  translated, for example "Vue Task".

When the call is performed to method `message_post`, the sessioon
context is not added to the RPC request. Therefore, when the model name
is read (method `_message_notification_recipients`), it is done in the
default language.

Note that this is the case for any RPC request using the `call` JS
function.

opw-707678