[FIX] mrp_repair: map_tax missing attributes
When the function map_tax is called on the fiscal position in repair
lines to compute taxes, the partner and product are not passed to the
method which leads to mistakes when the module account_taxcloud tries to
get back the taxes for the specific partner and tic_category of the
product.

To fix the issue, we pass the product and partner that we already have
to the function map_tax.